requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.11.1
django-psdb-engine==1.0.6
djangorestframework==3.14.0
mysqlclient==2.2.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
